// src/entities/activity-action.enum.ts
export enum ActivityAction {
  TICKET_CREATED = 'TICKET_CREATED',
  TITLE_UPDATED = 'TITLE_UPDATED',
  DESCRIPTION_UPDATED = 'DESCRIPTION_UPDATED',
  STATUS_CHANGED = 'STATUS_CHANGED',
  PRIORITY_CHANGED = 'PRIORITY_CHANGED',
  ASSIGNEE_CHANGED = 'ASSIGNEE_CHANGED',
  MOVED_TO_SPRINT = 'MOVED_TO_SPRINT',
  REMOVED_FROM_SPRINT = 'REMOVED_FROM_SPRINT',
  SPRINT_ASSIGNMENT_CHANGED = 'SPRINT_ASSIGNMENT_CHANGED',
  MOVED_TO_BACKLOG = 'MOVED_TO_BACKLOG',
  SPRINT_CHANGED = 'SPRINT_CHANGED',
}